Move randomization to its own module
